+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - plot: fix `bars` option
   - project tree: clear filter input when loading a session
   - server: fix serving files from paths containing accents
+  - editor: keep editor open when loading a session
 
 - widgets
   - image: add pre defined values for `size`, `position` and `repeat` properties